requirements.txt
asgiref==3.7.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==41.0.1
defusedxml==0.7.1
Django==4.0
django-allauth==0.54.0
django-environ==0.10.0
djangorestframework==3.14.0
idna==3.4
oauthlib==3.2.2
psycopg2-binary==2.9.2
pycparser==2.21
PyJWT==2.7.0
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
typing_extensions==4.7.0
urllib3==2.0.3
